backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.61.0
psycopg[binary]==3.2.3
pgvector==0.3.6